using new base class for tests renamed BotTestCase to TestBot
